@@ -7,7 +7,7 @@
     "prepublish": "babel src --out-dir lib --source-maps inline --ignore \"**/*.spec.js,**/__fixtures__,**/__mocks__\"",
     "test": "jest",
     "build": "yarn prepublish && webpack",
-    "start": "webpack-dev-server",
+    "start": "NODE_ENV=development webpack serve",
     "lint": "eslint .",
     "lint-and-fixup": "eslint --fix ."
   },
@@ -33,42 +33,46 @@
     "front-matter": "^3.0.0",
     "hastml": "^0.1.0",
     "he": "^1.1.1",
-    "highlight.js": "^9.6.0",
-    "jsesc": "^2.2.0",
-    "loader-utils": "^1.1.0",
-    "markdown-it": "^9.0.0",
+    "highlight.js": "^10.0.0",
+    "jsesc": "^3.0.0",
+    "loader-utils": "^2.0.0",
+    "markdown-it": "^11.0.0",
     "sha.js": "^2.4.5"
   },
   "devDependencies": {
-    "babel-cli": "6.26.0",
-    "babel-core": "6.26.3",
-    "babel-eslint": "10.0.3",
-    "babel-jest": "21.2.0",
-    "babel-loader": "8.0.6",
-    "babel-plugin-transform-object-rest-spread": "6.26.0",
+    "@babel/cli": "7.12.10",
+    "@babel/core": "7.12.10",
+    "@babel/plugin-proposal-object-rest-spread": "7.12.1",
+    "@babel/polyfill": "7.12.1",
+    "@babel/preset-env": "7.12.11",
+    "@babel/preset-react": "7.12.10",
+    "babel-core": "7.0.0-bridge.0",
+    "babel-eslint": "10.1.0",
+    "babel-jest": "23.6.0",
+    "babel-loader": "8.2.2",
     "babel-plugin-version-inline": "1.0.0",
-    "babel-polyfill": "6.26.0",
-    "babel-preset-es2015": "6.24.1",
-    "babel-preset-react": "6.24.1",
-    "codemirror": "5.48.4",
-    "css-loader": "3.2.0",
-    "eslint": "6.3.0",
-    "eslint-plugin-react": "7.5.1",
-    "extract-text-webpack-plugin": "3.0.2",
-    "file-loader": "4.2.0",
+    "codemirror": "5.59.1",
+    "css-loader": "3.6.0",
+    "css-minimizer-webpack-plugin": "1.1.5",
+    "eslint": "7.17.0",
+    "eslint-plugin-react": "7.22.0",
+    "file-loader": "6.0.0",
+    "html-webpack-plugin": "4.5.0",
     "image-webpack-loader": "6.0.0",
-    "jest": "21.2.1",
+    "jest": "25.5.4",
     "json-loader": "0.5.7",
-    "markdown-it-anchor": "4.0.0",
+    "markdown-it-anchor": "5.3.0",
     "markdown-it-table-of-contents": "0.4.4",
+    "mini-css-extract-plugin": "1.2.1",
     "prop-types": "15.7.2",
-    "react": "16.9.0",
-    "react-codemirror2": "6.0.0",
-    "react-dom": "16.9.0",
-    "react-test-renderer": "16.9.0",
-    "uglifyjs-webpack-plugin": "2.2.0",
-    "webpack": "4.39.3",
-    "webpack-dev-server": "3.8.0"
+    "react": "16.14.0",
+    "react-codemirror2": "7.2.1",
+    "react-dom": "16.14.0",
+    "react-test-renderer": "16.14.0",
+    "terser-webpack-plugin": "4.2.3",
+    "webpack": "4.43.0",
+    "webpack-cli": "4.1.0",
+    "webpack-dev-server": "3.11.0"
   },
   "peerDependencies": {
     "prop-types": "^15.5.10",
@@ -77,11 +81,11 @@
   "babel": {
     "plugins": [
       "version-inline",
-      "transform-object-rest-spread"
+      "@babel/plugin-proposal-object-rest-spread"
     ],
     "presets": [
-      "es2015",
-      "react"
+      "@babel/preset-env",
+      "@babel/preset-react"
     ]
   },
   "jest": {
