@@ -1,2 +1,3 @@
-webpackJsonp([1],{170:function(e,t,n){n(114),e.exports=n(372)},372:function(e,t,n){"use strict";var a=u(n(50)),r=u(n(149)),l=u(n(383));function u(e){return e&&e.__esModule?e:{default:e}}n(387),r.default.render(a.default.createElement(l.default,null),document.getElementById("root"))},383:function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0});var a=l(n(50)),r=l(n(384));function l(e){return e&&e.__esModule?e:{default:e}}function u(e){var t=e.className,r=e.style;return a.default.createElement("div",{className:t,style:r},a.default.createElement("center",null,a.default.createElement("h1",null,a.default.createElement("img",{id:"logo",src:n(156),alt:"Markdown Component Loader"})),a.default.createElement("p",null,"A Webpack loader that turns Markdown into dynamic, stateless React components!"),a.default.createElement("p",null,a.default.createElement("a",{className:"bubble-link github-link",href:"https://github.com/ticky/markdown-component-loader"},"Check it out on GitHub"))),a.default.createElement("h3",{id:"welcomemdx"},a.default.createElement("a",{className:"header-anchor",href:"#welcomemdx","aria-hidden":"true"},"🔗")," welcome.mdx"),a.default.createElement("pre",null,a.default.createElement("code",{className:"language-markdown"},a.default.createElement("span",{className:"emphasis"},"Hey there **","{{ props.who }}","** ✌🏼"),a.default.createElement("br",null))),a.default.createElement("h3",{id:"appjsx"},a.default.createElement("a",{className:"header-anchor",href:"#appjsx","aria-hidden":"true"},"🔗")," app.jsx"),a.default.createElement("pre",null,a.default.createElement("code",{className:"language-jsx"},a.default.createElement("span",{className:"emphasis"},"import Welcome from './welcome.mdx';"),a.default.createElement("br",null)," ",a.default.createElement("br",null),"ReactDOM.render(",a.default.createElement("br",null),a.default.createElement("span",{className:"emphasis"},'  <Welcome who="'+(e.who||"Monkey Magic")+'" />,'),a.default.createElement("br",null),"  document.getElementById('app')",a.default.createElement("br",null),");",a.default.createElement("br",null))),a.default.createElement("h3",{id:"rendered"},a.default.createElement("a",{className:"header-anchor",href:"#rendered","aria-hidden":"true"},"🔗")," Rendered:"),a.default.createElement("pre",null,a.default.createElement("code",{className:"language-html"},a.default.createElement("span",{className:"hljs-tag"},"<",a.default.createElement("span",{className:"hljs-name"},"div")," ",a.default.createElement("span",{className:"hljs-attr"},"id"),"=",a.default.createElement("span",{className:"hljs-string"},'"app"'),">"),a.default.createElement("br",null),a.default.createElement("span",{className:"emphasis"},"  <p>Hey there <strong>"+(e.who||"Monkey Magic")+"</strong> ✌🏼</p>"),a.default.createElement("br",null),a.default.createElement("span",{className:"hljs-tag"},"</",a.default.createElement("span",{className:"hljs-name"},"div"),">"),a.default.createElement("br",null))),a.default.createElement("h2",{id:"what-else-can-it-do"},a.default.createElement("a",{className:"header-anchor",href:"#what-else-can-it-do","aria-hidden":"true"},"🔗")," What else can it do?"),a.default.createElement("p",null,"Use JSX within Markdown. Plus import objects such as shared snippets, fancy library functions or even whole React components, using YAML front-matter:"),a.default.createElement("pre",null,a.default.createElement("code",{className:"language-markdown"},"---",a.default.createElement("br",null),a.default.createElement("span",{className:"emphasis"},"imports:"),a.default.createElement("br",null),a.default.createElement("span",{className:"emphasis"},"  numberFormat: number-format"),a.default.createElement("br",null),a.default.createElement("span",{className:"hljs-section"},a.default.createElement("span",{className:"emphasis"},"  Button: ui/button"),a.default.createElement("br",null),"---"),a.default.createElement("br",null),a.default.createElement("br",null),"You are visitor number ",a.default.createElement("span",{className:"emphasis"},"{{ numberFormat(1000000) }}"),"!",a.default.createElement("br",null),a.default.createElement("br",null),a.default.createElement("span",{className:"emphasis"},'<Button path="/browser-info">Details</Button>'),a.default.createElement("br",null))),a.default.createElement("p",null,"Destructured imports, with some added single quotes:"),a.default.createElement("pre",null,a.default.createElement("code",{className:"language-markdown"},"---",a.default.createElement("br",null),"imports:",a.default.createElement("br",null),a.default.createElement("span",{className:"hljs-section"},a.default.createElement("span",{className:"emphasis"},"  '{ window, document }': global"),a.default.createElement("br",null),"---"),a.default.createElement("br",null),a.default.createElement("br",null),"Hello, your user agent is ",a.default.createElement("span",{className:"emphasis"},"{{ window.navigator.userAgent }}"),".",a.default.createElement("br",null))),a.default.createElement("h2",{id:"give-it-a-go"},a.default.createElement("a",{className:"header-anchor",href:"#give-it-a-go","aria-hidden":"true"},"🔗")," Give it a go!"),a.default.createElement("p",null,"Want to see exactly what Markdown Component Loader will generate when it translates the mdx into a stateless React component?"),a.default.createElement("div",{className:"repl-image"},a.default.createElement("div",null,a.default.createElement("a",{className:"bubble-link white-bubble",href:"repl.html"},"Launch editor"))),a.default.createElement("h2",{id:"get-started"},a.default.createElement("a",{className:"header-anchor",href:"#get-started","aria-hidden":"true"},"🔗")," Get started"),a.default.createElement("p",null,"Check out the ",a.default.createElement("a",{href:"https://github.com/ticky/markdown-component-loader"},"GitHub Readme")," for how to get started, including details on how to add styles and event handlers to the DOM nodes generated by the markdown, and other options for customizing the loader."),a.default.createElement("center",null,a.default.createElement("a",{className:"bubble-link github-link",href:"https://github.com/ticky/markdown-component-loader"},"Check it out on GitHub"),a.default.createElement("hr",null),a.default.createElement("p",null,"Made with 💜 by ",a.default.createElement("a",{href:"https://twitter.com/ticky"},"@ticky"))))}u.propTypes={className:r.default.string,style:r.default.object},u.title="Markdown Component Loader",t.default=u},384:function(e,t,n){(function(t){if("production"!==t.env.NODE_ENV){var a="function"==typeof Symbol&&Symbol.for&&Symbol.for("react.element")||60103;e.exports=n(385)(function(e){return"object"==typeof e&&null!==e&&e.$$typeof===a},!0)}else e.exports=n(386)()}).call(t,n(20))},385:function(e,t,n){"use strict";(function(t){var a=n(34),r=n(59),l=n(77),u=n(58),o=n(106),c=n(105);e.exports=function(e,n){var s="function"==typeof Symbol&&Symbol.iterator,i="@@iterator";var d="<<anonymous>>",f={array:y("array"),bool:y("boolean"),func:y("function"),number:y("number"),object:y("object"),string:y("string"),symbol:y("symbol"),any:h(a.thatReturnsNull),arrayOf:function(e){return h(function(t,n,a,r,l){if("function"!=typeof e)return new p("Property `"+l+"` of component `"+a+"` has invalid PropType notation inside arrayOf.");var u=t[n];if(!Array.isArray(u)){var c=E(u);return new p("Invalid "+r+" `"+l+"` of type `"+c+"` supplied to `"+a+"`, expected an array.")}for(var s=0;s<u.length;s++){var i=e(u,s,a,r,l+"["+s+"]",o);if(i instanceof Error)return i}return null})},element:function(){return h(function(t,n,a,r,l){var u=t[n];if(!e(u)){var o=E(u);return new p("Invalid "+r+" `"+l+"` of type `"+o+"` supplied to `"+a+"`, expected a single ReactElement.")}return null})}(),instanceOf:function(e){return h(function(t,n,a,r,l){if(!(t[n]instanceof e)){var u=e.name||d,o=function(e){if(!e.constructor||!e.constructor.name)return d;return e.constructor.name}(t[n]);return new p("Invalid "+r+" `"+l+"` of type `"+o+"` supplied to `"+a+"`, expected instance of `"+u+"`.")}return null})},node:function(){return h(function(e,t,n,a,r){if(!b(e[t]))return new p("Invalid "+a+" `"+r+"` supplied to `"+n+"`, expected a ReactNode.");return null})}(),objectOf:function(e){return h(function(t,n,a,r,l){if("function"!=typeof e)return new p("Property `"+l+"` of component `"+a+"` has invalid PropType notation inside objectOf.");var u=t[n],c=E(u);if("object"!==c)return new p("Invalid "+r+" `"+l+"` of type `"+c+"` supplied to `"+a+"`, expected an object.");for(var s in u)if(u.hasOwnProperty(s)){var i=e(u,s,a,r,l+"."+s,o);if(i instanceof Error)return i}return null})},oneOf:function(e){if(!Array.isArray(e))return"production"!==t.env.NODE_ENV&&l(!1,"Invalid argument supplied to oneOf, expected an instance of array."),a.thatReturnsNull;return h(function(t,n,a,r,l){for(var u=t[n],o=0;o<e.length;o++)if(m(u,e[o]))return null;var c=JSON.stringify(e);return new p("Invalid "+r+" `"+l+"` of value `"+u+"` supplied to `"+a+"`, expected one of "+c+".")})},oneOfType:function(e){if(!Array.isArray(e))return"production"!==t.env.NODE_ENV&&l(!1,"Invalid argument supplied to oneOfType, expected an instance of array."),a.thatReturnsNull;for(var n=0;n<e.length;n++){var r=e[n];if("function"!=typeof r)return l(!1,"Invalid argument supplied to oneOfType. Expected an array of check functions, but received %s at index %s.",g(r),n),a.thatReturnsNull}return h(function(t,n,a,r,l){for(var u=0;u<e.length;u++){var c=e[u];if(null==c(t,n,a,r,l,o))return null}return new p("Invalid "+r+" `"+l+"` supplied to `"+a+"`.")})},shape:function(e){return h(function(t,n,a,r,l){var u=t[n],c=E(u);if("object"!==c)return new p("Invalid "+r+" `"+l+"` of type `"+c+"` supplied to `"+a+"`, expected `object`.");for(var s in e){var i=e[s];if(i){var d=i(u,s,a,r,l+"."+s,o);if(d)return d}}return null})},exact:function(e){return h(function(t,n,a,r,l){var c=t[n],s=E(c);if("object"!==s)return new p("Invalid "+r+" `"+l+"` of type `"+s+"` supplied to `"+a+"`, expected `object`.");var i=u({},t[n],e);for(var d in i){var f=e[d];if(!f)return new p("Invalid "+r+" `"+l+"` key `"+d+"` supplied to `"+a+"`.\nBad object: "+JSON.stringify(t[n],null,"  ")+"\nValid keys: "+JSON.stringify(Object.keys(e),null,"  "));var m=f(c,d,a,r,l+"."+d,o);if(m)return m}return null})}};function m(e,t){return e===t?0!==e||1/e==1/t:e!=e&&t!=t}function p(e){this.message=e,this.stack=""}function h(e){if("production"!==t.env.NODE_ENV)var a={},u=0;function c(c,s,i,f,m,h,y){if(f=f||d,h=h||i,y!==o)if(n)r(!1,"Calling PropTypes validators directly is not supported by the `prop-types` package. Use `PropTypes.checkPropTypes()` to call them. Read more at http://fb.me/use-check-prop-types");else if("production"!==t.env.NODE_ENV&&"undefined"!=typeof console){var b=f+":"+i;!a[b]&&u<3&&(l(!1,"You are manually calling a React.PropTypes validation function for the `%s` prop on `%s`. This is deprecated and will throw in the standalone `prop-types` package. You may be seeing this warning due to a third-party PropTypes library. See https://fb.me/react-warning-dont-call-proptypes for details.",h,f),a[b]=!0,u++)}return null==s[i]?c?null===s[i]?new p("The "+m+" `"+h+"` is marked as required in `"+f+"`, but its value is `null`."):new p("The "+m+" `"+h+"` is marked as required in `"+f+"`, but its value is `undefined`."):null:e(s,i,f,m,h)}var s=c.bind(null,!1);return s.isRequired=c.bind(null,!0),s}function y(e){return h(function(t,n,a,r,l,u){var o=t[n];return E(o)!==e?new p("Invalid "+r+" `"+l+"` of type `"+v(o)+"` supplied to `"+a+"`, expected `"+e+"`."):null})}function b(t){switch(typeof t){case"number":case"string":case"undefined":return!0;case"boolean":return!t;case"object":if(Array.isArray(t))return t.every(b);if(null===t||e(t))return!0;var n=function(e){var t=e&&(s&&e[s]||e[i]);if("function"==typeof t)return t}(t);if(!n)return!1;var a,r=n.call(t);if(n!==t.entries){for(;!(a=r.next()).done;)if(!b(a.value))return!1}else for(;!(a=r.next()).done;){var l=a.value;if(l&&!b(l[1]))return!1}return!0;default:return!1}}function E(e){var t=typeof e;return Array.isArray(e)?"array":e instanceof RegExp?"object":function(e,t){return"symbol"===e||"Symbol"===t["@@toStringTag"]||"function"==typeof Symbol&&t instanceof Symbol}(t,e)?"symbol":t}function v(e){if(void 0===e||null===e)return""+e;var t=E(e);if("object"===t){if(e instanceof Date)return"date";if(e instanceof RegExp)return"regexp"}return t}function g(e){var t=v(e);switch(t){case"array":case"object":return"an "+t;case"boolean":case"date":case"regexp":return"a "+t;default:return t}}return p.prototype=Error.prototype,f.checkPropTypes=c,f.PropTypes=f,f}}).call(t,n(20))},386:function(e,t,n){"use strict";var a=n(34),r=n(59),l=n(106);e.exports=function(){function e(e,t,n,a,u,o){o!==l&&r(!1,"Calling PropTypes validators directly is not supported by the `prop-types` package. Use PropTypes.checkPropTypes() to call them. Read more at http://fb.me/use-check-prop-types")}function t(){return e}e.isRequired=e;var n={array:e,bool:e,func:e,number:e,object:e,string:e,symbol:e,any:e,arrayOf:t,element:e,instanceOf:t,node:e,objectOf:t,oneOf:t,oneOfType:t,shape:t,exact:t};return n.checkPropTypes=a,n.PropTypes=n,n}},387:function(e,t){}},[170]);
+/*! For license information please see site.js.LICENSE.txt */
+!function(e){function t(t){for(var r,l,u=t[0],c=t[1],i=t[2],f=0,p=[];f<u.length;f++)l=u[f],Object.prototype.hasOwnProperty.call(a,l)&&a[l]&&p.push(a[l][0]),a[l]=0;for(r in c)Object.prototype.hasOwnProperty.call(c,r)&&(e[r]=c[r]);for(s&&s(t);p.length;)p.shift()();return o.push.apply(o,i||[]),n()}function n(){for(var e,t=0;t<o.length;t++){for(var n=o[t],r=!0,u=1;u<n.length;u++){var c=n[u];0!==a[c]&&(r=!1)}r&&(o.splice(t--,1),e=l(l.s=n[0]))}return e}var r={},a={3:0},o=[];function l(t){if(r[t])return r[t].exports;var n=r[t]={i:t,l:!1,exports:{}};return e[t].call(n.exports,n,n.exports,l),n.l=!0,n.exports}l.m=e,l.c=r,l.d=function(e,t,n){l.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:n})},l.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},l.t=function(e,t){if(1&t&&(e=l(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var n=Object.create(null);if(l.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var r in e)l.d(n,r,function(t){return e[t]}.bind(null,r));return n},l.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return l.d(t,"a",t),t},l.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},l.p="";var u=window.webpackJsonp=window.webpackJsonp||[],c=u.push.bind(u);u.push=t,u=u.slice();for(var i=0;i<u.length;i++)t(u[i]);var s=c;o.push([332,0]),n()}({0:function(e,t,n){"use strict";e.exports=n(15)},11:function(e,t,n){"use strict";var r=Object.getOwnPropertySymbols,a=Object.prototype.hasOwnProperty,o=Object.prototype.propertyIsEnumerable;function l(e){if(null==e)throw new TypeError("Object.assign cannot be called with null or undefined");return Object(e)}e.exports=function(){try{if(!Object.assign)return!1;var e=new String("abc");if(e[5]="de","5"===Object.getOwnPropertyNames(e)[0])return!1;for(var t={},n=0;n<10;n++)t["_"+String.fromCharCode(n)]=n;if("0123456789"!==Object.getOwnPropertyNames(t).map((function(e){return t[e]})).join(""))return!1;var r={};return"abcdefghijklmnopqrst".split("").forEach((function(e){r[e]=e})),"abcdefghijklmnopqrst"===Object.keys(Object.assign({},r)).join("")}catch(e){return!1}}()?Object.assign:function(e,t){for(var n,u,c=l(e),i=1;i<arguments.length;i++){for(var s in n=Object(arguments[i]))a.call(n,s)&&(c[s]=n[s]);if(r){u=r(n);for(var f=0;f<u.length;f++)o.call(n,u[f])&&(c[u[f]]=n[u[f]])}}return c}},15:function(e,t,n){"use strict";var r=n(11),a="function"==typeof Symbol&&Symbol.for,o=a?Symbol.for("react.element"):60103,l=a?Symbol.for("react.portal"):60106,u=a?Symbol.for("react.fragment"):60107,c=a?Symbol.for("react.strict_mode"):60108,i=a?Symbol.for("react.profiler"):60114,s=a?Symbol.for("react.provider"):60109,f=a?Symbol.for("react.context"):60110,p=a?Symbol.for("react.forward_ref"):60112,m=a?Symbol.for("react.suspense"):60113,d=a?Symbol.for("react.memo"):60115,h=a?Symbol.for("react.lazy"):60116,y="function"==typeof Symbol&&Symbol.iterator;function b(e){for(var t="https://reactjs.org/docs/error-decoder.html?invariant="+e,n=1;n<arguments.length;n++)t+="&args[]="+encodeURIComponent(arguments[n]);return"Minified React error #"+e+"; visit "+t+" for the full message or use the non-minified dev environment for full errors and additional helpful warnings."}var v={isMounted:function(){return!1},enqueueForceUpdate:function(){},enqueueReplaceState:function(){},enqueueSetState:function(){}},E={};function w(e,t,n){this.props=e,this.context=t,this.refs=E,this.updater=n||v}function g(){}function _(e,t,n){this.props=e,this.context=t,this.refs=E,this.updater=n||v}w.prototype.isReactComponent={},w.prototype.setState=function(e,t){if("object"!=typeof e&&"function"!=typeof e&&null!=e)throw Error(b(85));this.updater.enqueueSetState(this,e,t,"setState")},w.prototype.forceUpdate=function(e){this.updater.enqueueForceUpdate(this,e,"forceUpdate")},g.prototype=w.prototype;var k=_.prototype=new g;k.constructor=_,r(k,w.prototype),k.isPureReactComponent=!0;var O={current:null},j=Object.prototype.hasOwnProperty,x={key:!0,ref:!0,__self:!0,__source:!0};function S(e,t,n){var r,a={},l=null,u=null;if(null!=t)for(r in void 0!==t.ref&&(u=t.ref),void 0!==t.key&&(l=""+t.key),t)j.call(t,r)&&!x.hasOwnProperty(r)&&(a[r]=t[r]);var c=arguments.length-2;if(1===c)a.children=n;else if(1<c){for(var i=Array(c),s=0;s<c;s++)i[s]=arguments[s+2];a.children=i}if(e&&e.defaultProps)for(r in c=e.defaultProps)void 0===a[r]&&(a[r]=c[r]);return{$$typeof:o,type:e,key:l,ref:u,props:a,_owner:O.current}}function N(e){return"object"==typeof e&&null!==e&&e.$$typeof===o}var C=/\/+/g,P=[];function T(e,t,n,r){if(P.length){var a=P.pop();return a.result=e,a.keyPrefix=t,a.func=n,a.context=r,a.count=0,a}return{result:e,keyPrefix:t,func:n,context:r,count:0}}function R(e){e.result=null,e.keyPrefix=null,e.func=null,e.context=null,e.count=0,10>P.length&&P.push(e)}function M(e,t,n,r){var a=typeof e;"undefined"!==a&&"boolean"!==a||(e=null);var u=!1;if(null===e)u=!0;else switch(a){case"string":case"number":u=!0;break;case"object":switch(e.$$typeof){case o:case l:u=!0}}if(u)return n(r,e,""===t?"."+L(e,0):t),1;if(u=0,t=""===t?".":t+":",Array.isArray(e))for(var c=0;c<e.length;c++){var i=t+L(a=e[c],c);u+=M(a,i,n,r)}else if(null===e||"object"!=typeof e?i=null:i="function"==typeof(i=y&&e[y]||e["@@iterator"])?i:null,"function"==typeof i)for(e=i.call(e),c=0;!(a=e.next()).done;)u+=M(a=a.value,i=t+L(a,c++),n,r);else if("object"===a)throw n=""+e,Error(b(31,"[object Object]"===n?"object with keys {"+Object.keys(e).join(", ")+"}":n,""));return u}function A(e,t,n){return null==e?0:M(e,"",t,n)}function L(e,t){return"object"==typeof e&&null!==e&&null!=e.key?function(e){var t={"=":"=0",":":"=2"};return"$"+(""+e).replace(/[=:]/g,(function(e){return t[e]}))}(e.key):t.toString(36)}function $(e,t){e.func.call(e.context,t,e.count++)}function I(e,t,n){var r=e.result,a=e.keyPrefix;e=e.func.call(e.context,t,e.count++),Array.isArray(e)?D(e,r,n,(function(e){return e})):null!=e&&(N(e)&&(e=function(e,t){return{$$typeof:o,type:e.type,key:t,ref:e.ref,props:e.props,_owner:e._owner}}(e,a+(!e.key||t&&t.key===e.key?"":(""+e.key).replace(C,"$&/")+"/")+n)),r.push(e))}function D(e,t,n,r,a){var o="";null!=n&&(o=(""+n).replace(C,"$&/")+"/"),A(e,I,t=T(t,o,r,a)),R(t)}var F={current:null};function B(){var e=F.current;if(null===e)throw Error(b(321));return e}var q={ReactCurrentDispatcher:F,ReactCurrentBatchConfig:{suspense:null},ReactCurrentOwner:O,IsSomeRendererActing:{current:!1},assign:r};t.Children={map:function(e,t,n){if(null==e)return e;var r=[];return D(e,r,null,t,n),r},forEach:function(e,t,n){if(null==e)return e;A(e,$,t=T(null,null,t,n)),R(t)},count:function(e){return A(e,(function(){return null}),null)},toArray:function(e){var t=[];return D(e,t,null,(function(e){return e})),t},only:function(e){if(!N(e))throw Error(b(143));return e}},t.Component=w,t.Fragment=u,t.Profiler=i,t.PureComponent=_,t.StrictMode=c,t.Suspense=m,t.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED=q,t.cloneElement=function(e,t,n){if(null==e)throw Error(b(267,e));var a=r({},e.props),l=e.key,u=e.ref,c=e._owner;if(null!=t){if(void 0!==t.ref&&(u=t.ref,c=O.current),void 0!==t.key&&(l=""+t.key),e.type&&e.type.defaultProps)var i=e.type.defaultProps;for(s in t)j.call(t,s)&&!x.hasOwnProperty(s)&&(a[s]=void 0===t[s]&&void 0!==i?i[s]:t[s])}var s=arguments.length-2;if(1===s)a.children=n;else if(1<s){i=Array(s);for(var f=0;f<s;f++)i[f]=arguments[f+2];a.children=i}return{$$typeof:o,type:e.type,key:l,ref:u,props:a,_owner:c}},t.createContext=function(e,t){return void 0===t&&(t=null),(e={$$typeof:f,_calculateChangedBits:t,_currentValue:e,_currentValue2:e,_threadCount:0,Provider:null,Consumer:null}).Provider={$$typeof:s,_context:e},e.Consumer=e},t.createElement=S,t.createFactory=function(e){var t=S.bind(null,e);return t.type=e,t},t.createRef=function(){return{current:null}},t.forwardRef=function(e){return{$$typeof:p,render:e}},t.isValidElement=N,t.lazy=function(e){return{$$typeof:h,_ctor:e,_status:-1,_result:null}},t.memo=function(e,t){return{$$typeof:d,type:e,compare:void 0===t?null:t}},t.useCallback=function(e,t){return B().useCallback(e,t)},t.useContext=function(e,t){return B().useContext(e,t)},t.useDebugValue=function(){},t.useEffect=function(e,t){return B().useEffect(e,t)},t.useImperativeHandle=function(e,t,n){return B().useImperativeHandle(e,t,n)},t.useLayoutEffect=function(e,t){return B().useLayoutEffect(e,t)},t.useMemo=function(e,t){return B().useMemo(e,t)},t.useReducer=function(e,t,n){return B().useReducer(e,t,n)},t.useRef=function(e){return B().useRef(e)},t.useState=function(e){return B().useState(e)},t.version="16.13.1"},16:function(e,t,n){"use strict";var r,a,o,l,u;if("undefined"==typeof window||"function"!=typeof MessageChannel){var c=null,i=null,s=function(){if(null!==c)try{var e=t.unstable_now();c(!0,e),c=null}catch(e){throw setTimeout(s,0),e}},f=Date.now();t.unstable_now=function(){return Date.now()-f},r=function(e){null!==c?setTimeout(r,0,e):(c=e,setTimeout(s,0))},a=function(e,t){i=setTimeout(e,t)},o=function(){clearTimeout(i)},l=function(){return!1},u=t.unstable_forceFrameRate=function(){}}else{var p=window.performance,m=window.Date,d=window.setTimeout,h=window.clearTimeout;if("undefined"!=typeof console){var y=window.cancelAnimationFrame;"function"!=typeof window.requestAnimationFrame&&console.error("This browser doesn't support requestAnimationFrame. Make sure that you load a polyfill in older browsers. https://fb.me/react-polyfills"),"function"!=typeof y&&console.error("This browser doesn't support cancelAnimationFrame. Make sure that you load a polyfill in older browsers. https://fb.me/react-polyfills")}if("object"==typeof p&&"function"==typeof p.now)t.unstable_now=function(){return p.now()};else{var b=m.now();t.unstable_now=function(){return m.now()-b}}var v=!1,E=null,w=-1,g=5,_=0;l=function(){return t.unstable_now()>=_},u=function(){},t.unstable_forceFrameRate=function(e){0>e||125<e?console.error("forceFrameRate takes a positive int between 0 and 125, forcing framerates higher than 125 fps is not unsupported"):g=0<e?Math.floor(1e3/e):5};var k=new MessageChannel,O=k.port2;k.port1.onmessage=function(){if(null!==E){var e=t.unstable_now();_=e+g;try{E(!0,e)?O.postMessage(null):(v=!1,E=null)}catch(e){throw O.postMessage(null),e}}else v=!1},r=function(e){E=e,v||(v=!0,O.postMessage(null))},a=function(e,n){w=d((function(){e(t.unstable_now())}),n)},o=function(){h(w),w=-1}}function j(e,t){var n=e.length;e.push(t);e:for(;;){var r=n-1>>>1,a=e[r];if(!(void 0!==a&&0<N(a,t)))break e;e[r]=t,e[n]=a,n=r}}function x(e){return void 0===(e=e[0])?null:e}function S(e){var t=e[0];if(void 0!==t){var n=e.pop();if(n!==t){e[0]=n;e:for(var r=0,a=e.length;r<a;){var o=2*(r+1)-1,l=e[o],u=o+1,c=e[u];if(void 0!==l&&0>N(l,n))void 0!==c&&0>N(c,l)?(e[r]=c,e[u]=n,r=u):(e[r]=l,e[o]=n,r=o);else{if(!(void 0!==c&&0>N(c,n)))break e;e[r]=c,e[u]=n,r=u}}}return t}return null}function N(e,t){var n=e.sortIndex-t.sortIndex;return 0!==n?n:e.id-t.id}var C=[],P=[],T=1,R=null,M=3,A=!1,L=!1,$=!1;function I(e){for(var t=x(P);null!==t;){if(null===t.callback)S(P);else{if(!(t.startTime<=e))break;S(P),t.sortIndex=t.expirationTime,j(C,t)}t=x(P)}}function D(e){if($=!1,I(e),!L)if(null!==x(C))L=!0,r(F);else{var t=x(P);null!==t&&a(D,t.startTime-e)}}function F(e,n){L=!1,$&&($=!1,o()),A=!0;var r=M;try{for(I(n),R=x(C);null!==R&&(!(R.expirationTime>n)||e&&!l());){var u=R.callback;if(null!==u){R.callback=null,M=R.priorityLevel;var c=u(R.expirationTime<=n);n=t.unstable_now(),"function"==typeof c?R.callback=c:R===x(C)&&S(C),I(n)}else S(C);R=x(C)}if(null!==R)var i=!0;else{var s=x(P);null!==s&&a(D,s.startTime-n),i=!1}return i}finally{R=null,M=r,A=!1}}function B(e){switch(e){case 1:return-1;case 2:return 250;case 5:return 1073741823;case 4:return 1e4;default:return 5e3}}var q=u;t.unstable_IdlePriority=5,t.unstable_ImmediatePriority=1,t.unstable_LowPriority=4,t.unstable_NormalPriority=3,t.unstable_Profiling=null,t.unstable_UserBlockingPriority=2,t.unstable_cancelCallback=function(e){e.callback=null},t.unstable_continueExecution=function(){L||A||(L=!0,r(F))},t.unstable_getCurrentPriorityLevel=function(){return M},t.unstable_getFirstCallbackNode=function(){return x(C)},t.unstable_next=function(e){switch(M){case 1:case 2:case 3:var t=3;break;default:t=M}var n=M;M=t;try{return e()}finally{M=n}},t.unstable_pauseExecution=function(){},t.unstable_requestPaint=q,t.unstable_runWithPriority=function(e,t){switch(e){case 1:case 2:case 3:case 4:case 5:break;default:e=3}var n=M;M=e;try{return t()}finally{M=n}},t.unstable_scheduleCallback=function(e,n,l){var u=t.unstable_now();if("object"==typeof l&&null!==l){var c=l.delay;c="number"==typeof c&&0<c?u+c:u,l="number"==typeof l.timeout?l.timeout:B(e)}else l=B(e),c=u;return e={id:T++,callback:n,priorityLevel:e,startTime:c,expirationTime:l=c+l,sortIndex:-1},c>u?(e.sortIndex=c,j(P,e),null===x(C)&&e===x(P)&&($?o():$=!0,a(D,c-u))):(e.sortIndex=l,j(C,e),L||A||(L=!0,r(F))),e},t.unstable_shouldYield=function(){var e=t.unstable_now();I(e);var n=x(C);return n!==R&&null!==R&&null!==n&&null!==n.callback&&n.startTime<=e&&n.expirationTime<R.expirationTime||l()},t.unstable_wrapCallback=function(e){var t=M;return function(){var n=M;M=t;try{return e.apply(this,arguments)}finally{M=n}}}},17:function(e,t,n){"use strict";n.r(t),t.default=n.p+"10505e11e3e9a645e3b5c46dbf1e1f28.svg"},23:function(e,t,n){e.exports=n(42)()},27:function(e,t,n){"use strict";e.exports=n(16)},332:function(e,t,n){"use strict";n.r(t);var r=n(0),a=n.n(r),o=n(9),l=n.n(o),u=n(23),c=n.n(u);function i(e){var t=e.className,r=e.style;return a.a.createElement("div",{className:t,style:r},a.a.createElement("center",null,a.a.createElement("h1",null,a.a.createElement("img",{id:"logo",src:n(17).default,alt:"Markdown Component Loader"})),a.a.createElement("p",null,"A Webpack loader that turns Markdown into dynamic, stateless React components!"),a.a.createElement("p",null,a.a.createElement("a",{className:"bubble-link github-link",href:"https://github.com/ticky/markdown-component-loader"},"Check it out on GitHub"))),a.a.createElement("h3",{id:"welcome.mdx"},a.a.createElement("a",{className:"header-anchor",href:"#welcome.mdx"},"🔗")," welcome.mdx"),a.a.createElement("pre",null,a.a.createElement("code",{className:"language-markdown"},a.a.createElement("span",{className:"emphasis"},"Hey there **","{{ props.who }}","** ✌🏼"),a.a.createElement("br",null))),a.a.createElement("h3",{id:"app.jsx"},a.a.createElement("a",{className:"header-anchor",href:"#app.jsx"},"🔗")," app.jsx"),a.a.createElement("pre",null,a.a.createElement("code",{className:"language-jsx"},a.a.createElement("span",{className:"emphasis"},"import Welcome from './welcome.mdx';"),a.a.createElement("br",null)," ",a.a.createElement("br",null),"ReactDOM.render(",a.a.createElement("br",null),a.a.createElement("span",{className:"emphasis"},'  <Welcome who="'.concat(e.who||"Monkey Magic",'" />,')),a.a.createElement("br",null),"  document.getElementById('app')",a.a.createElement("br",null),");",a.a.createElement("br",null))),a.a.createElement("h3",{id:"rendered%3A"},a.a.createElement("a",{className:"header-anchor",href:"#rendered%3A"},"🔗")," Rendered:"),a.a.createElement("pre",null,a.a.createElement("code",{className:"language-html"},a.a.createElement("span",{className:"hljs-tag"},"<",a.a.createElement("span",{className:"hljs-name"},"div")," ",a.a.createElement("span",{className:"hljs-attr"},"id"),"=",a.a.createElement("span",{className:"hljs-string"},'"app"'),">"),a.a.createElement("br",null),a.a.createElement("span",{className:"emphasis"},"  <p>Hey there <strong>".concat(e.who||"Monkey Magic","</strong> ✌🏼</p>")),a.a.createElement("br",null),a.a.createElement("span",{className:"hljs-tag"},"</",a.a.createElement("span",{className:"hljs-name"},"div"),">"),a.a.createElement("br",null))),a.a.createElement("h2",{id:"what-else-can-it-do%3F"},a.a.createElement("a",{className:"header-anchor",href:"#what-else-can-it-do%3F"},"🔗")," What else can it do?"),a.a.createElement("p",null,"Use JSX within Markdown. Plus import objects such as shared snippets, fancy library functions or even whole React components, using YAML front-matter:"),a.a.createElement("pre",null,a.a.createElement("code",{className:"language-markdown"},"---",a.a.createElement("br",null),a.a.createElement("span",{className:"emphasis"},"imports:"),a.a.createElement("br",null),a.a.createElement("span",{className:"emphasis"},"  numberFormat: number-format"),a.a.createElement("br",null),a.a.createElement("span",{className:"hljs-section"},a.a.createElement("span",{className:"emphasis"},"  Button: ui/button"),a.a.createElement("br",null),"---"),a.a.createElement("br",null),a.a.createElement("br",null),"You are visitor number ",a.a.createElement("span",{className:"emphasis"},"{{ numberFormat(1000000) }}"),"!",a.a.createElement("br",null),a.a.createElement("br",null),a.a.createElement("span",{className:"emphasis"},'<Button path="/browser-info">Details</Button>'),a.a.createElement("br",null))),a.a.createElement("p",null,"Destructured imports, with some added single quotes:"),a.a.createElement("pre",null,a.a.createElement("code",{className:"language-markdown"},"---",a.a.createElement("br",null),"imports:",a.a.createElement("br",null),a.a.createElement("span",{className:"hljs-section"},a.a.createElement("span",{className:"emphasis"},"  '{ window, document }': global"),a.a.createElement("br",null),"---"),a.a.createElement("br",null),a.a.createElement("br",null),"Hello, your user agent is ",a.a.createElement("span",{className:"emphasis"},"{{ window.navigator.userAgent }}"),".",a.a.createElement("br",null))),a.a.createElement("h2",{id:"give-it-a-go!"},a.a.createElement("a",{className:"header-anchor",href:"#give-it-a-go!"},"🔗")," Give it a go!"),a.a.createElement("p",null,"Want to see exactly what Markdown Component Loader will generate when it translates the mdx into a stateless React component?"),a.a.createElement("div",{className:"repl-image"},a.a.createElement("div",null,a.a.createElement("a",{className:"bubble-link white-bubble",href:"repl.html"},"Launch editor"))),a.a.createElement("h2",{id:"get-started"},a.a.createElement("a",{className:"header-anchor",href:"#get-started"},"🔗")," Get started"),a.a.createElement("p",null,"Check out the ",a.a.createElement("a",{href:"https://github.com/ticky/markdown-component-loader"},"GitHub Readme")," for how to get started, including details on how to add styles and event handlers to the DOM nodes generated by the markdown, and other options for customizing the loader."),a.a.createElement("center",null,a.a.createElement("a",{className:"bubble-link github-link",href:"https://github.com/ticky/markdown-component-loader"},"Check it out on GitHub"),a.a.createElement("hr",null),a.a.createElement("p",null,"Made with 💜 by ",a.a.createElement("a",{href:"https://twitter.com/ticky"},"@ticky"))))}i.propTypes={className:c.a.string,style:c.a.object},i.title="Markdown Component Loader";var s=i;l.a.render(a.a.createElement(s,null),document.getElementById("root"))},42:function(e,t,n){"use strict";var r=n(43);function a(){}function o(){}o.resetWarningCache=a,e.exports=function(){function e(e,t,n,a,o,l){if(l!==r){var u=new Error("Calling PropTypes validators directly is not supported by the `prop-types` package. Use PropTypes.checkPropTypes() to call them. Read more at http://fb.me/use-check-prop-types");throw u.name="Invariant Violation",u}}function t(){return e}e.isRequired=e;var n={array:e,bool:e,func:e,number:e,object:e,string:e,symbol:e,any:e,arrayOf:t,element:e,elementType:e,instanceOf:t,node:e,objectOf:t,oneOf:t,oneOfType:t,shape:t,exact:t,checkPropTypes:o,resetWarningCache:a};return n.PropTypes=n,n}},43:function(e,t,n){"use strict";e.exports="SECRET_DO_NOT_PASS_THIS_OR_YOU_WILL_BE_FIRED"},9:function(e,t,n){"use strict";!function e(){if("undefined"!=typeof __REACT_DEVTOOLS_GLOBAL_HOOK__&&"function"==typeof __REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE)try{__REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE(e)}catch(e){console.error(e)}}(),e.exports=n(26)}});
 //# sourceMappingURL=site.js.map
